Enroute to #25. Lanscape orientation change.
NPE appears to be occuring during a landscape mode change (from
portrait). There are other examples of this happening with a
LunarLanding example that's included in the SDK examples.

Could be from the (NexusS) simulator being slow with the CPU cycles.
Emulates a much slower machine and is more likely to stumble upon this.