For #28. Begin networked physics architecture.
There are three modes we want to support:

1. Singleplayer
2. Multiplayer
3. Adhoc

Singelplayer is self-explanatory, with everything running locally.

Multiplayer is a client-server architecture where the-server-is-king.
This suggests we'll need to create our own protocol rather than some
webapp RESTful interface. Additionally, we need prediction on the client
side so that the UI will be updated immediately. When the server returns
with updated and differentiating results, roll back to The Server's
Word and reapply (ie. predict) the user behavior--involves a circular
buffer of activity.

Adhoc wil be peer to peer (wifi direct, bluetooth, etc.) which will
not be ranked play, but also not have a server involved. This peer
to peer behavior will require a different set of threads to be run.